requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
google-api-core==2.7.1
google-api-python-client==2.47.0
google-auth==2.6.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
numpy==1.22.3
oauth2client==4.1.3
pandas==1.4.2
pkg_resources==0.0.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
pyTelegramBotAPI==4.5.0
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rsa==4.8
schedule==1.1.0
six==1.16.0
telebot==0.0.4
telegram==0.0.1
tornado==6.1
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
